Delete/ADD: Izdzēsu nevajadzīgos failus un pievienoju Services prieks Thesis
