--- src/main/java/lv/venta/services/ITestCRUDService.java
@@ -1,5 +1,0 @@
-package lv.venta.services;
-
-public interface ITestCRUDService {
-
-}
