+++ src/main/java/lv/venta/services/IThesisCRUDService.java
@@ -1,0 +1,5 @@
+package lv.venta.services;
+
+public interface IThesisCRUDService {
+
+}
