+++ src/main/java/lv/venta/services/ThesisCRUDService.java
@@ -1,0 +1,8 @@
+package lv.venta.services;
+
+import org.springframework.stereotype.Service;
+
+@Service
+public class ThesisCRUDService implements IThesisCRUDService{
+
+}
